refactor: encapsulate hostRulesFromEnv
