+++ lib/workers/global/config/parse/env.ts
@@ -3,9 +3,8 @@ import is from '@sindresorhus/is';
 import { getOptions } from '../../../../config/options';
 import type { AllConfig, RenovateOptions } from '../../../../config/types';
 import { PlatformId } from '../../../../constants';
-import { getDatasourceList } from '../../../../datasource';
 import { logger } from '../../../../logger';
-import type { HostRule } from '../../../../types';
+import { hostRulesFromEnv } from './host-rules-from-env';
 
 function normalizePrefixes(
   env: NodeJS.ProcessEnv,
@@ -120,58 +119,10 @@ export function getConfig(inputEnv: NodeJS.ProcessEnv): AllConfig {
     });
   }
 
-  const datasources = new Set(getDatasourceList());
-  const fields = ['token', 'username', 'password'];
-
-  const hostRules: HostRule[] = [];
-
-  const npmEnvPrefixes = ['npm_config_', 'npm_lifecycle_', 'npm_package_'];
-
-  for (const envName of Object.keys(env).sort()) {
-    if (npmEnvPrefixes.some((prefix) => envName.startsWith(prefix))) {
-      logger.trace('Ignoring npm env: ' + envName);
-      continue; // eslint-disable-line no-continue
-    }
-    // Double underscore __ is used in place of hyphen -
-    const splitEnv = envName.toLowerCase().replace(/__/g, '-').split('_');
-    const hostType = splitEnv.shift();
-    if (datasources.has(hostType)) {
-      const suffix = splitEnv.pop();
-      if (fields.includes(suffix)) {
-        let matchHost: string;
-        const rule: HostRule = {};
-        rule[suffix] = env[envName];
-        if (splitEnv.length === 0) {
-          // host-less rule
-        } else if (splitEnv.length === 1) {
-          logger.warn(`Cannot parse ${envName} env`);
-          continue; // eslint-disable-line no-continue
-        } else {
-          matchHost = splitEnv.join('.');
-        }
-        const existingRule = hostRules.find(
-          (hr) => hr.hostType === hostType && hr.matchHost === matchHost
-        );
-        logger.debug(`Converting ${envName} into a global host rule`);
-        if (existingRule) {
-          // Add current field to existing rule
-          existingRule[suffix] = env[envName];
-        } else {
-          // Create a new rule
-          const newRule: HostRule = {
-            hostType,
-          };
-          if (matchHost) {
-            newRule.matchHost = matchHost;
-          }
-          newRule[suffix] = env[envName];
-          hostRules.push(newRule);
-        }
-      }
-    }
+  const hostRules = hostRulesFromEnv(env);
+  if (hostRules) {
+    config.hostRules = [...config.hostRules, ...hostRules];
   }
-
-  config.hostRules = [...config.hostRules, ...hostRules];
 
   // These env vars are deprecated and deleted to make sure they're not used
   const unsupportedEnv = [
